requirements.txt
# Flask==3.0.3
# Flask-SocketIO==5.3.6
# eventlet==0.36.1
# opencv-python==4.10.0.84
# mediapipe==0.10.14
# tensorflow-macos==2.15.0
# python-dotenv==1.0.1
# numpy==1.26.4

flask
flask-cors
flask-socketio
python-dotenv
pymongo[srv]
firebase-admin
tensorflow
opencv-python-headless
mediapipe
numpy
gevent
gevent-websocket
eventlet
